test: move csv_to_db unit tests to test_conversion
